backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-dotenv==1.0.0
slowapi==0.1.9
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib==1.2.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.14.0
prophet==1.1.5
statsmodels==0.14.0
joblib==1.3.2
aiofiles==23.2.1
schedule==1.2.0
schedule==1.2.0
email-validator==2.1.0
aiofiles==23.2.1
PyYAML==6.0.1
requests==2.31.0